+++ src/components/Products/Products.js
@@ -1,34 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import ProductCards from './ProductCard'
-
-
-const products = [
-    {
-        id: 1,
-        name: 'Produto legal',
-        price: 123,
-        photo: 'https://picsum.photos/200/200?a=1'
-    },
-    {
-        id: 2,
-        name: 'Produto 2',
-        price: 200,
-        photo: 'https://picsum.photos/200/200?a=2'
-    },
-    {
-        id: 3,
-        name: 'Produto 3',
-        price: 30,
-        photo: 'https://picsum.photos/200/200?a=3'
-    },
-    {
-        id: 4,
-        name: 'Produto 4',
-        price: 10,
-        photo: 'https://picsum.photos/200/200?a=4'
-    }
-]
+import ProductCards from './ProductCard';
 
 const ProductsContainer = styled.div`
     display:flex;
@@ -60,7 +32,7 @@ class Products extends React.Component {
         return (
             <ProductsContainer>
                 <ProductsHeader>
-                    <p>Quantidade de produtos: {products.length}</p>
+                    <p>Quantidade de produtos: {this.props.products.length}</p>
                     <ProductsOrder>
                         <p>Ordenação: </p>
                         <select>
@@ -71,7 +43,7 @@ class Products extends React.Component {
                 </ProductsHeader>
 
                 <ProductCards 
-                    products={products}
+                    products={this.props.products}
                 />
             </ProductsContainer>
         );
